requirements.txt
kaleido
cohere
langchain-community 
langchainhub 
langchain-openai
openai
tiktoken
typing-extensions == 4.7.1
unstructured[all-docs]
langchain
PyPDF2
faiss-cpu
safetensors
huggingface-hub
InstructorEmbedding
sentence-transformers
torch
transformers
ctransformers
chromadb
unstructured
pdf2image
pdfminer.six
pypdf
python-docx
chromadb
Flask
flask_cors
line-bot-sdk
googletrans-py